LayoutTests/accessibility/mac/selected-visible-position-range-expected.txt

First paragraph.

Last paragraph.

This tests getting and setting the selected VisiblePosition range.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS current selection is 'Last paragraph.'
PASS current selection is 'First paragraph.

'
PASS current selection is ''
PASS current selection index is 19
PASS successfullyParsed is true

TEST COMPLETE
